requirements.txt
kivy==2.3.1
Pillow==10.4.0
bcrypt==4.2.0
